+++ bubblegum/core/planner/intent.py
@@ -37,6 +37,7 @@ def build_options(
         "stability_timeout_ms",
         "stability_spinner_selectors",
         "select_no_filter",
+        "fallback_selector",
     }
     opts = {k: v for k, v in kwargs.items() if k in known}
     opts.setdefault("use_ai", ai_enabled)